refactor: extract WalletManager accessors and error types
Improves the key-wallet-manager crate structure by breaking up lib.rs:

- New accessors.rs (~213 lines): wallet creation, import, removal,
  and lookup methods extracted from WalletManager
- New error.rs (~99 lines): WalletError enum extracted into its own
  module
- Slimmed lib.rs: now contains only the core WalletManager struct
  definition and field management
